Autostart functionality for all difficulties
The prior autostart used the file difficulty to start the timer. It was a byte equal to 1 for Easy, 2 for Normal, and so on. Its default value prior to starting a new game is 2, and the timer checked != 2 to see if a run was starting. As a consequence of that, the timer couldn't autostart on Normal difficulty.

The game sets IGT to 0 at that same moment when a new game is begun, so we can just use that fact to start the timer on any difficulty. (If this ever breaks for anyone, although I don't think it should, the fix is to check if current.fileTimer is less than some small value, rather than equal to 0.)

I should note that I've assumed the IGT's location in memory for the "AS" and "community" versions of the game, based on its location for the other two versions. In the two versions that I've checked, the IGT is located 8 bytes ahead of the "diffStart" byte that was used previously, and I don't anticipate that being different elsewhere.